moved all spotify related endpoints into a separate router
