+++ Backend/package.json
@@ -9,6 +9,7 @@
     "express-flash": "^0.0.2",
     "express-session": "^1.18.0",
     "jsonwebtoken": "^9.0.2",
+    "node-fetch": "^3.3.2",
     "nodemon": "^3.1.4",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0"
